moved SubstrateHeadersSyncPipeline to headers_pipeline.rs
