+++ relays/substrate/src/headers_target.rs
@@ -18,47 +18,20 @@
 //! runtime that implements `<BridgedChainName>HeaderApi` to allow bridging with
 //! <BridgedName> chain.
 
+use crate::headers_pipeline::SubstrateHeadersSyncPipeline;
+
 use async_trait::async_trait;
 use codec::{Decode, Encode};
 use futures::TryFutureExt;
 use headers_relay::{
 	sync_loop::TargetClient,
-	sync_types::{HeaderIdOf, HeadersSyncPipeline, QueuedHeader, SubmittedHeaders},
+	sync_types::{HeaderIdOf, QueuedHeader, SubmittedHeaders},
 };
 use relay_substrate_client::{Chain, Client, Error as SubstrateError};
 use relay_utils::HeaderId;
 use sp_core::Bytes;
 use sp_runtime::Justification;
 use std::collections::HashSet;
-
-/// Headers sync pipeline for Substrate <-> Substrate relays.
-#[async_trait]
-pub trait SubstrateHeadersSyncPipeline: HeadersSyncPipeline {
-	/// Name of the `best_block` runtime method.
-	const BEST_BLOCK_METHOD: &'static str;
-	/// Name of the `finalized_block` runtime method.
-	const FINALIZED_BLOCK_METHOD: &'static str;
-	/// Name of the `is_known_block` runtime method.
-	const IS_KNOWN_BLOCK_METHOD: &'static str;
-	/// Name of the `incomplete_headers` runtime method.
-	const INCOMPLETE_HEADERS_METHOD: &'static str;
-
-	/// Signed transaction type.
-	type SignedTransaction: Send + Sync + Encode;
-
-	/// Make submit header transaction.
-	async fn make_submit_header_transaction(
-		&self,
-		header: QueuedHeader<Self>,
-	) -> Result<Self::SignedTransaction, SubstrateError>;
-
-	/// Make completion transaction for the header.
-	async fn make_complete_header_transaction(
-		&self,
-		id: HeaderIdOf<Self>,
-		completion: Justification,
-	) -> Result<Self::SignedTransaction, SubstrateError>;
-}
 
 /// Substrate client as Substrate headers target.
 pub struct SubstrateHeadersTarget<C: Chain, P> {
